moved folders to enhance the lib structure
