+++ src/design/organism/slide/slide.css
@@ -1,4 +1,4 @@
-@value colors: "../../style/colors.css";
+@value colors: "../../variables/colors.css";
 @value secondary from colors;
 @value defaultColor: secondary;
