+++ src/renderer/virtual-dom.js
